commit dimenticato

+++ src/main/java/org/galvanica/service/BagnoService.java
@@ -30,33 +30,44 @@ public class BagnoService implements ICRUDService<Bagno, BagnoDto> {
         }
         if (elemento.getScattiTotali() == null)
             elemento.setScattiTotali(0);
-        if (elemento.getScattiParziali() == null)
-            elemento.setScattiParziali(0);
+        if (elemento.getRestoScatti() == null)
+            elemento.setRestoScatti(0);
         Bagno bagno = Bagno.builder()
                 .nome(elemento.getNome())
                 .litri(elemento.getLitri())
                 .scattiTotali(elemento.getScattiTotali())
-                .scattiParziali(elemento.getScattiParziali())
+                .restoScatti(elemento.getRestoScatti())
                 .build();
         return repository.save(bagno);
     }
 
     @Override
     public Boolean elimina(long id) {
-        repository.deleteById(id);
-        return repository.findById(id).isEmpty();
-
+        boolean risposta = repository.findById(id).isPresent();
+        if (risposta) {
+            repository.deleteById(id);
+        }
+        return risposta;
     }
 
     @Override
-    public Bagno aggiorna(BagnoDto elemento) {
-        return null;
-        //repository.save(elemento);
+    public Bagno aggiorna(BagnoDto elemento, long id) {
+        Optional<Bagno> bagno = repository.findById(id);
+        if (bagno.isEmpty()) {
+            throw new RuntimeException(
+                    "metti un id corretto che questo non c'è, bischero.");
+        }
+        Bagno b = bagno.get();
+        b.setNome(elemento.getNome());
+        b.setRestoScatti(elemento.getRestoScatti());
+        b.setScattiTotali(elemento.getScattiTotali());
+        b.setLitri(elemento.getLitri());
+
+        return repository.save(b);
     }
 
     @Override
     public Optional<Bagno> ricercaId(long id) {
-        return null;
-        //repository.findById(id);
+        return repository.findById(id);
     }
 }
